requirements.txt
pytorch-ignite==0.4.2
tqdm==4.48.0
scikit-learn==0.23.1
torch~=1.5.1
numpy~=1.19.0
